+++ src/lib/builders/slider/create.ts
@@ -15,12 +15,17 @@ import {
 	overridable,
 	styleToString,
 	toWritableStores,
+	add,
+	div,
+	mul,
+	sub,
+	clamp,
 } from '$lib/internal/helpers/index.js';
 import type { MeltActionReturn } from '$lib/internal/types.js';
 import { derived, get, writable } from 'svelte/store';
 import { generateIds } from '../../internal/helpers/id';
 import type { SliderEvents } from './events.js';
-import { add, div, mul, sub } from './helpers.js';
+
 import type { CreateSliderProps } from './types.js';
 
 const defaults = {
@@ -49,74 +54,7 @@ export const createSlider = (props?: CreateSliderProps) => {
 
 	const meltIds = generateIds(['root']);
 
-	const root = builder(name(), {
-		stores: [disabled, orientation],
-		returned: ([$disabled, $orientation]) => {
-			return {
-				disabled: disabledAttr($disabled),
-				'aria-disabled': ariaDisabledAttr($disabled),
-				'data-orientation': $orientation,
-				style: $disabled ? undefined : 'touch-action: none;',
-				'data-melt-id': meltIds.root,
-			};
-		},
-	});
-
-	const position = derived([min, max], ([$min, $max]) => {
-		return (val: number) => {
-			const pos = mul(div(sub(val, $min), sub($max, $min)), 100);
-			return pos;
-		};
-	});
-
-	const ticks = derived([min, max, step], ([$min, $max, $step]) => {
-		const difference = sub($max, $min);
-
-		// min = 0, max = 8, step = 3:
-		// ----------------------------
-		// 0, 3, 6
-		// (8 - 0) / 3 = 2.666... = 3 ceiled
-		let count = Math.ceil(div(difference, $step));
-
-		// min = 0, max = 9, step = 3:
-		// ---------------------------
-		// 0, 3, 6, 9
-		// (9 - 0) / 3 = 3
-		// We need to add 1 because `difference` is a multiple of `step`.
-		if (difference % $step == 0) {
-			count++;
-		}
-
-		return count;
-	});
-
-	const actualMax = derived([min, step, ticks], ([$min, $step, $ticks]) => {
-		const numberOfSteps = $ticks - 1;
-		// Actual max value numberOfSteps multiplied by step and added to min.
-		const actualMax = add($min, mul(numberOfSteps, $step));
-		return actualMax;
-	});
-
-	const range = builder(name('range'), {
-		stores: [value, orientation, position],
-		returned: ([$value, $orientation, $position]) => {
-			const minimum = $value.length > 1 ? $position(Math.min(...$value) ?? 0) : 0;
-			const maximum = 100 - $position(Math.max(...$value) ?? 0);
-
-			const orientationStyles =
-				$orientation === 'horizontal'
-					? { left: `${minimum}%`, right: `${maximum}%` }
-					: { top: `${maximum}%`, bottom: `${minimum}%` };
-
-			return {
-				style: styleToString({
-					position: 'absolute',
-					...orientationStyles,
-				}),
-			};
-		},
-	});
-
+	// Helpers
 	const updatePosition = (val: number, index: number) => {
 		value.update((prev) => {
 			if (!prev) return [val];
@@ -155,6 +93,82 @@ export const createSlider = (props?: CreateSliderProps) => {
 			(thumb): thumb is HTMLElement => isHTMLElement(thumb)
 		);
 	};
+
+	// States
+	const position = derived([min, max], ([$min, $max]) => {
+		return (val: number) => {
+			const pos = mul(div(sub(val, $min), sub($max, $min)), 100);
+			return pos;
+		};
+	});
+
+	const ticks = derived([min, max, step], ([$min, $max, $step]) => {
+		const difference = sub($max, $min);
+
+		// min = 0, max = 8, step = 3:
+		// ----------------------------
+		// 0, 3, 6
+		// (8 - 0) / 3 = 2.666... = 3 ceiled
+		let count = Math.ceil(div(difference, $step));
+
+		// min = 0, max = 9, step = 3:
+		// ---------------------------
+		// 0, 3, 6, 9
+		// (9 - 0) / 3 = 3
+		// We need to add 1 because `difference` is a multiple of `step`.
+		if (difference % $step == 0) {
+			count++;
+		}
+
+		return count;
+	});
+
+	/**
+	 * Represents the actual max value of the slider, taking into account the
+	 * number of steps and the step value.
+	 * e.g. given a min of 0, a max of 10, and a step of 3, the actual max value
+	 * would be 9.
+	 *
+	 */
+	const actualMax = derived([min, step, ticks], ([$min, $step, $ticks]) => {
+		const numberOfSteps = $ticks - 1;
+		// Actual max value numberOfSteps multiplied by step and added to min.
+		return add($min, mul(numberOfSteps, $step));
+	});
+
+	// Elements
+	const root = builder(name(), {
+		stores: [disabled, orientation],
+		returned: ([$disabled, $orientation]) => {
+			return {
+				disabled: disabledAttr($disabled),
+				'aria-disabled': ariaDisabledAttr($disabled),
+				'data-orientation': $orientation,
+				style: $disabled ? undefined : 'touch-action: none;',
+				'data-melt-id': meltIds.root,
+			};
+		},
+	});
+
+	const range = builder(name('range'), {
+		stores: [value, orientation, position],
+		returned: ([$value, $orientation, $position]) => {
+			const minimum = $value.length > 1 ? $position(Math.min(...$value) ?? 0) : 0;
+			const maximum = 100 - $position(Math.max(...$value) ?? 0);
+
+			const orientationStyles =
+				$orientation === 'horizontal'
+					? { left: `${minimum}%`, right: `${maximum}%` }
+					: { top: `${maximum}%`, bottom: `${minimum}%` };
+
+			return {
+				style: styleToString({
+					position: 'absolute',
+					...orientationStyles,
+				}),
+			};
+		},
+	});
 
 	const thumb = builder(name('thumb'), {
 		stores: [value, position, min, max, disabled, orientation],
@@ -331,6 +345,7 @@ export const createSlider = (props?: CreateSliderProps) => {
 		},
 	});
 
+	// Effects
 	effect(
 		[root, min, max, actualMax, disabled, orientation, step],
 		([$root, $min, $max, $actualMax, $disabled, $orientation, $step]) => {
@@ -447,6 +462,27 @@ export const createSlider = (props?: CreateSliderProps) => {
 			};
 		}
 	);
+
+	effect([step, min, max, value], function fixValue([$step, $min, $max, $value]) {
+		const isValidValue = (v: number) => {
+			if (v < $min || v > $max) return false;
+			if (sub(v, $min) % $step !== 0) return false;
+			return true;
+		};
+
+		if ($value.some((v) => !isValidValue(v))) {
+			value.update((prev) => {
+				return [...prev].map((v) => {
+					if (isValidValue(v)) return v;
+					const withoutMin = sub(v, $min);
+					const stepNum = Math.floor(div(withoutMin, $step));
+
+					const newValue = add($min, mul(stepNum, $step));
+					return clamp($min, newValue, $max);
+				});
+			});
+		}
+	});
 
 	return {
 		ids: meltIds,
