+++ src/lib/internal/helpers/index.ts
@@ -24,3 +24,4 @@ export * from './typeahead.js';
 export * from './elements.js';
 export * from './ignore.js';
 export * from './focus.js';
+export * from './math.js';
